fixed error
git-svn-id: http://streamboard.de.vu/svn/oscam/trunk@10827 4b0bc96b-bc66-0410-9d44-ebda105a78c1

+++ oscam-garbage.h
@@ -2,7 +2,7 @@
 #define OSCAM_GARBAGE_H_
 
 #ifdef WITH_DEBUG
-extern void add_garbage_debug(void *data, char *file, uint16_t line);
+extern void add_garbage_debug(void *data, char *file, uint32_t line);
 #define add_garbage(x) add_garbage_debug(x, __FILE__, __LINE__)
 #else
 extern void add_garbage(void *data);
